+++ Kernel/GenericInterface/Operation/Ticket/TicketUpdate.pm
@@ -253,6 +253,10 @@ sub Run {
         );
     }
 
+    if ( $UserType eq 'Customer' ) {
+        $UserID = $Self->{ConfigObject}->Get('CustomerPanelUserID')
+    }
+
     # check TicketID
     my $TicketID;
     if ( $Param{Data}->{TicketNumber} ) {
@@ -1408,7 +1412,7 @@ sub _TicketUpdate {
     my %TicketData = $Self->{TicketObject}->TicketGet(
         TicketID      => $TicketID,
         DynamicFields => 0,
-        UserID        => $Param{UserId},
+        UserID        => $Param{UserID},
     );
 
     # update ticket parameters